requirements.txt
PyYAML~=6.0.2rc1
SQLAlchemy~=2.0.31
typeguard~=4.3.0
click~=8.1.7